# Conflicts:
#	.idea/deviceManager.xml

+++ app/src/main/java/com/example/projekmobile_kel11/data/model/User.kt
@@ -5,10 +5,11 @@ data class User(
     var nama: String = "",
     var email: String = "",
     var role: String = "",
-    var usia: Int = 0,
+    var usia: Long = 0L,   // 🔥 FIX
     var gender: String = "",
     var phone: String = "",
     var photoUrl: String = "",
     var createdAt: Long = 0L,
     var updatedAt: Long = 0L
 )
+
